refactor(Drawer): re-org file-structure
