+++ src/containers/tool/Drawer/styles/header/close_line.js
@@ -1,30 +1,9 @@
 import styled from 'styled-components'
 
 // see example: https://codepen.io/mattbraun/pen/EywBJR
-// import { css } from '@/utils'
+// import { css, theme } from '@/utils'
 
-const BaseWrapper = styled.div`
-  z-index: 1;
-  position: absolute;
-  bottom: 10px;
-  left: 0;
-  width: 100%;
-  height: 30px;
-  background: #002a35;
-  border-bottom: 1px solid;
-  border-bottom-color: #144150;
-  border-bottom-left-radius: 16px;
-  border-bottom-right-radius: 16px;
-`
-export const BottomWrapper = styled(BaseWrapper)`
-  top: 10px;
-  transform: rotate(180deg);
-`
-export const TopWrapper = styled(BaseWrapper)`
-  bottom: 10px;
-`
-
-export const CloseLineWrapper = styled.div`
+export const Wrapper = styled.div`
   position: relative;
   transform: rotate(180deg);
   margin-right: calc(50% - 19px);
